creacion de crrview

+++ PMMX.Modelo/Entidades/Desperdicio.cs
@@ -11,7 +11,7 @@ namespace PMMX.Modelo.Entidades
     public class Desperdicio
     {
         public int Id { get; set; }
-        public int Cantidad { get; set; }
+        public Double Cantidad { get; set; }
         public DateTime Fecha { get; set; }
         public int IdPersona { get; set; }
         public int IdWorkCenter { get; set; }
